added dashboard route to public blueprint
