+++ src/crweb/web.py
@@ -173,69 +173,6 @@ def teardown_request(exception):
         pass
 
 
-# Dashboard
-
-
-# Dashboard Home
-@app.route('/dashboard')
-def dashboard_page():
-    ''' Dashboard: Generate the Welcome/Status page for the dashboard '''
-    verify = verifyLogin(
-        app.config['SECRET_KEY'], app.config['COOKIE_TIMEOUT'], request.cookies)
-    if verify:
-        user = User()
-        user.get('uid', verify, g.rdb_conn)
-        data = startData(user)
-        data['active'] = 'dashboard'
-        data['url'] = '/dashboard'
-        data['js_bottom'].append("screen-o-death.js")
-        data['js_bottom'].append("screen-o-death-chart.js")
-        if user.status != "active":
-            data['url'] = '/dashboard/mod-subscription'
-            page = render_template('mod-subscription.html', data=data)
-        else:
-
-            data['monitors'] = user.getMonitors(g.rdb_conn)
-            data['reactions'] = user.getReactions(g.rdb_conn)
-            data['monevents'] = user.getEvents(g.rdb_conn)
-            data['moneventsnum'] = len(data['monevents'])
-            data['monstats'] = {'healthy': 0,
-                                'unknown': 0,
-                                'failed': 0}
-            for key in data['monitors'].keys():
-                if "healthy" in data['monitors'][key]['status']:
-                    data['monstats']['healthy'] = data[
-                        'monstats']['healthy'] + 1
-                elif "failed" in data['monitors'][key]['status']:
-                    data['monstats']['failed'] = data['monstats']['failed'] + 1
-                else:
-                    data['monstats']['unknown'] = data[
-                        'monstats']['unknown'] + 1
-
-            # If there are no monitors print a welcome message
-            if len(data['monitors']) < 1 and len(data['reactions']) < 1:
-                data['welcome'] = True
-            else:
-                data['welcome'] = False
-
-            if len(data['monitors']) < 1:
-                data['mons'] = False
-            else:
-                data['mons'] = True
-
-            if len(data['reactions']) < 1:
-                data['reacts'] = False
-            else:
-                data['reacts'] = True
-
-            from generalforms import subscribe
-            form = subscribe.AddPackForm(request.form)
-            page = render_template('screen-o-death.html', data=data, form=form)
-        return page
-    else:
-        return redirect(url_for('user.login_page'))
-
-
 # Monitors
 
 
@@ -524,7 +461,7 @@ def checkaction_page(cid, action):
                 print("/dashboard/action-checks - Manual monitor action failed: do not own")
                 flash('It does not appear you own this health check', 'danger')
 
-    return redirect(url_for('dashboard_page'))
+    return redirect(url_for('public.dashboard_page'))
 
 
 # Delete Checks
@@ -553,7 +490,7 @@ def delcheck_page(cid):
             else:
                 print("/dashboard/delete-checks - Delete failed")
                 flash('Health Check was not deleted', 'danger')
-    return redirect(url_for('dashboard_page'))
+    return redirect(url_for('public.dashboard_page'))
 
 
 # Web Checks
@@ -937,7 +874,7 @@ def delreaction_page(rid):
                     flash('Reaction was not deleted', 'danger')
             else:
                 flash('You must first detach this reaction from all monitors before deleting', 'danger')
-    return redirect(url_for('dashboard_page'))
+    return redirect(url_for('public.dashboard_page'))
 
 
 # Mod-Subscription
